[refactor] laundryshop 코틀린 마이그레이션

+++ src/main/kotlin/org/washcode/washpang/domain/handledItems/repository/HandledItemsRepository.kt
@@ -14,10 +14,10 @@ import java.util.*
 interface HandledItemsRepository : JpaRepository<HandledItems, Long> {
     override fun findAll(): List<HandledItems> //모든 데이터 가져오기
 
-    //List<HandledItems> findByLaundryshopId(Long laundryShopId);
+
     //세탁소 아이디를 받아서 handledItem 내용 조회(id, laundryshop_id,price, item_name, category)
     @Query("SELECT h FROM HandledItems h WHERE h.laundryshop.id = :laundryshopId")
-    fun findByLaundryshopId(@Param("laundryshopId") laundryshopId: Long): List<HandledItems>
+    fun findByLaundryshopId(@Param("laundryshopId") laundryshopId: Int): List<HandledItems>
 
     //카테고리별로 세탁소 id 조회
     @Query("SELECT h.laundryshop.id FROM HandledItems h WHERE h.category = :category")
